+++ Blank/main.py
@@ -1,5 +1,5 @@
 """ 
- Copyright (C) 2018 Claude SIMON (http://q37.info/contact/).
+ 	Copyright (C) 2018 Claude SIMON (http://q37.info/contact/).
 
 	This file is part of XDHq.
 
@@ -19,8 +19,8 @@
 
 import os, sys
 
-if not "EPEIOS_SRV" in os.environ:
-	sys.path.append("Atlas.python.zip")
+sys.path.append("./Atlas.python.zip")
+sys.path.append("../Atlas.python.zip")
 
 import Atlas