+++ Hello/main.py
@@ -19,8 +19,8 @@
 
 import os, sys
 
-if not "EPEIOS_SRV" in os.environ:
-	sys.path.append("Atlas.python.zip")
+sys.path.append("./Atlas.python.zip")
+sys.path.append("../Atlas.python.zip")
 
 import Atlas